Neutron Reflector Graphite Sanity Reduction (#296)
Halves the amount of Graphite that is required, as even if you took the cheapest route, the Tungstencarbide recipe, will take a **minimum** of 256 Graphite for a single Iridium Neutron Reflector. Or in this case, if you wanted to make a fusion T2 or T3 reactor (with the one included for the reactor itself), it'll take 16896 Graphite.