Changing telepad recipe from dense americium plates to dense enriched naquadah plates
Reasoning for this is that i want to downtier the telepad from late zpm to mid zpm and to remove the bottleneck to crafting it since a single fusion mk 2 will take around 54 mins to craft a telepad.

However, once the player gets wetware circuits they can just make the far better relocator item instead.